# docs/enterprise-digital-twin/contracts/openapi/enterprise-digital-twin.openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Digital Twin API
  version: 1.0.0
  summary: Permission-aware graph, simulation, governed-action, and synthetic physical-asset API.
  description: Tenant context is always derived from the authenticated principal and a server-issued context handle. Tenant identifiers supplied in payloads are consistency assertions and never select authorization scope. Raw X-Tenant-ID is rejected.
...
servers:
  - url: https://api.enterprise-digital-twin.example
security:
  - oidc: [edt.read]
    contextHeader: []
  - oidc: [edt.read]
    contextCookie: []
...
paths:
...
  /v1/twin/prediction/models:
    post:
      tags: [AI Model Registry]
      operationId: registerPredictiveModel
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/ModelRegister'}}}}
      responses:
        '201': {description: Versioned non-LLM model definition., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    get:
      tags: [AI Model Registry]
      operationId: listPredictiveModels
      parameters:
        - {name: kind, in: query, schema: {type: string}}
        - {name: prediction_kind, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 2000}}
      responses:
        '200': {description: Bounded tenant model page., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/models/{model_id}:
    get:
      tags: [AI Model Registry]
      operationId: getPredictiveModel
      parameters: [{name: model_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Model inputs outputs version accuracy owner and trigger., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/runs:
    post:
      tags: [Predictive Intelligence]
      operationId: createPrediction
      description: Processes historical data through an active registered non-LLM model and returns forecast confidence and limitations.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/PredictionCreate'}}}}
      responses:
        '201': {description: Pending-outcome prediction., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/runs/{prediction_id}:
    get:
      tags: [Predictive Intelligence]
      operationId: getPrediction
      parameters: [{name: prediction_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Tenant-scoped prediction and validation state., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/runs/{prediction_id}/outcomes:
    post:
      tags: [Predictive Intelligence]
      operationId: recordPredictionOutcome
      parameters:
        - {name: prediction_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/OutcomeRecord'}}}}
      responses:
        '201': {description: Real observed outcome awaiting user validation., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/runs/{prediction_id}/validations:
    post:
      tags: [Predictive Intelligence]
      operationId: validatePredictionOutcome
      parameters:
        - {name: prediction_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/ValidationRecord'}}}}
      responses:
        '201': {description: Validation metrics plus updated model accuracy and learning revision., headers: {ETag: {schema: {type: string}}}, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/prediction/knowledge:
    post:
      tags: [Predictive Intelligence]
      operationId: submitPredictionKnowledge
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/decision-intelligence.schema.json#/$defs/KnowledgeSubmit'}}}}
      responses:
        '201': {description: Pending-review historical outcome specification rule correction or expert knowledge., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/connectors:
    get:
      tags: [Connectors]
      operationId: listConnectorInstallations
      security: [{oidc: [edt.connector.read], contextHeader: []}, {oidc: [edt.connector.read], contextCookie: []}]
      responses:
        '200': {description: Tenant connector health and scope., content: {application/json: {schema: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/ConnectorInstallation'}}}}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/twin/connectors:
    get:
      tags: [Twin Data Foundation]
      operationId: listTwinConnectorDefinitions
      x-edt-required-capabilities: [connector.admin]
      responses:
        '200': {description: Governed connector definitions without credential material., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Twin Data Foundation]
      operationId: createTwinConnectorDefinition
      description: Registers ERP, CRM, HRIS, accounting, API, database, or document integration policy metadata. Registration does not execute the connector.
      x-edt-required-capabilities: [connector.admin]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/data-foundation.schema.json#/$defs/connectorDefinition'}}}}
      responses:
        '201': {description: Registered connector definition and concurrency ETag., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/connectors/{connectorId}:
    get:
      tags: [Twin Data Foundation]
      operationId: getTwinConnectorDefinition
      x-edt-required-capabilities: [connector.admin]
      parameters: [{name: connectorId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Connector definition., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    patch:
      tags: [Twin Data Foundation]
      operationId: updateTwinConnectorDefinition
      x-edt-required-capabilities: [connector.admin]
      parameters:
        - {name: connectorId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody: {required: true, content: {application/json: {schema: {type: object, minProperties: 1}}}}
      responses:
        '200': {description: Updated connector definition., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/mcp-servers:
    get:
      tags: [Twin Data Foundation]
      operationId: listTwinMcpServerDefinitions
      x-edt-required-capabilities: [connector.admin]
      responses:
        '200': {description: Governed MCP server definitions., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Twin Data Foundation]
      operationId: createTwinMcpServerDefinition
      description: Registers MCP name, purpose, secret reference, tools, permissions, and connected data. Registration does not create a transport or execute tools.
      x-edt-required-capabilities: [connector.admin]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '../schemas/data-foundation.schema.json#/$defs/mcpServerDefinition'}}}}
      responses:
        '201': {description: Registered MCP definition and concurrency ETag., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /v1/twin/mcp-servers/{serverId}:
    get:
      tags: [Twin Data Foundation]
      operationId: getTwinMcpServerDefinition
      x-edt-required-capabilities: [connector.admin]
      parameters: [{name: serverId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: MCP server definition., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
    patch:
      tags: [Twin Data Foundation]
      operationId: updateTwinMcpServerDefinition
      x-edt-required-capabilities: [connector.admin]
      parameters:
        - {name: serverId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody: {required: true, content: {application/json: {schema: {type: object, minProperties: 1}}}}
      responses:
        '200': {description: Updated MCP server definition., content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.enterprise-digital-twin.example/.well-known/openid-configuration
    contextHeader:
      type: apiKey
      in: header
      name: X-EDT-Context
      description: Opaque short-lived audience-bound server-issued membership context handle. Raw tenant identifiers are rejected.
    contextCookie:
      type: apiKey
      in: cookie
      name: EDT-Context
      description: Browser-only HTTP-only Secure SameSite opaque context handle; CSRF controls apply to commands.
  parameters:
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string, minLength: 3, maxLength: 200}}
...
  responses:
    Problem:
      description: RFC 9457-compatible error with no unauthorized resource existence disclosure.
      content: {application/problem+json: {schema: {$ref: '../schemas/canonical.schema.json#/$defs/Problem'}}}
  schemas:
...
    ConnectorInstallation:
      type: object
      additionalProperties: false
      required: [installation_id, provider, status, scopes, last_reconciled_at, freshness_seconds]
      properties:
        installation_id: {type: string, format: uuid}
        provider: {enum: [github, jira]}
        status: {enum: [healthy, degraded, paused, revoked, error]}
        scopes: {type: array, items: {type: string}}
        last_reconciled_at: {type: string, format: date-time}
        freshness_seconds: {type: integer, minimum: 0}
...